enhance(view): auth-header
